+++ server/app.js
@@ -7,9 +7,10 @@ const express = require('express')
 const morgan = require('morgan')
 const helmet = require('helmet')
 
+
 const cors = require('cors')
 const spotifyRouter = require('./controllers/spotify')
-
+const generalRouter = require('./controllers/general')
 
 const app = express()
 console.log('mode: ', process.env.NODE_ENV)
@@ -29,47 +30,10 @@ app.use(morgan('tiny'))
 app.use(cors())
 app.use(helmet())
 
+
 app.use(express.static(path.resolve(__dirname, '../react-ui/build')))
 
-app.get('/home', (req, res) => {
-    console.log('hello from backend')
-    res.sendFile(path.resolve(__dirname, '../react-ui/build', 'index.html'))
-})
-
-app.get('/about', (req, res) => {
-    res.send('<h1>Hello World</h1>')
-})
-
-
+app.use('', generalRouter)
 app.use('/api/spotify', spotifyRouter)
-
-const resume = require('./resumeData.json')
-app.get('/resume', async (req, res) => {
-    return res.json(resume)
-})
-
-const Form = require('./models/form')
-app.post('/contact/api', (req, res) => {
-    const body = req.body
-
-    const form = {
-        ...body
-    }
-    Form.create(form) //could also use the async await syntax but would have to install express async errors lib
-        .then(savedForm => {
-            console.log(`Saved the form from ${form.name}`)
-            res.json(savedForm)
-        })
-        .catch(err => console.log(err))
-
-})
-
-app.get('/api', (req, res) => {
-    Form.find({})
-        .then(forms => {
-            res.json(forms)
-        })
-        .catch(err => console.log(err))
-})
 
 module.exports = app
