+++ server/resumeData.json
@@ -13,7 +13,7 @@
       "project":"https://github.com/AlexWang18/Personal-Website",
       "twitter":"http://twitter.com/AlexunderWang",
       "website": "alexwang18.github.io",
-      "resumedownload":"Alex Wang Resume.docx",
+      "resumedownload":"Alex-Wang-Resume.docx",
       "endingpic" : "snowbody.JPG",
       "social":[
         {
